+++ lib/actions.js
@@ -1,429 +1,89 @@
-/* Copyright (c) 2010-2017 Richard Rodger and other contributors, MIT License */
+/* Copyright (c) 2017 Richard Rodger and other contributors, MIT License */
 'use strict'
 
 
-var Util = require('util')
-var Assert = require('assert')
+// var Assert = require('assert')
 
 
 var _ = require('lodash')
-var Jsonic = require('jsonic')
 
 
 var Common = require('./common')
 
 
-// Find the action metadata for a given pattern, if it exists.
-exports.find = function (pattern, flags) {
-  var seneca = this
+module.exports = function (root) {
+  root.stats = make_action_seneca_stats(root.private$)
 
-  var pat = _.isString(pattern) ? Jsonic(pattern) : pattern
-  pat = seneca.util.clean(pat)
-  pat = pat || {}
+  // Add builtin actions.
+  root.add({role: 'seneca', cmd: 'stats'}, root.stats)
+  root.add({role: 'seneca', cmd: 'close'}, action_seneca_close)
+  root.add({role: 'seneca', info: 'fatal'}, action_seneca_fatal)
+  root.add({role: 'seneca', get: 'options'}, action_options_get)
 
-  var actmeta = seneca.private$.actrouter.find(pat)
-
-  if (!actmeta && flags && flags.catchall) {
-    actmeta = seneca.private$.actrouter.find({})
-  }
-
-  return actmeta
+  // Legacy builtin actions.
+  // Remove in Seneca 4.x
+  root.add({role: 'seneca', stats: true, deprecate$: true}, root.stats)
+  root.add({role: 'options', cmd: 'get', deprecate$: true}, action_options_get)
 }
 
 
-// True if an action matching the pattern exists.
-exports.has = function (pattern) {
-  return !!exports.find.call(this, pattern)
+function action_seneca_fatal (args, done) {
+  done()
 }
 
 
-// List all actions that match the pattern.
-exports.list = function (pattern) {
-  var pat = _.isString(pattern) ? Jsonic(pattern) : pattern
-  var found = this.private$.actrouter.list(pat)
-  found = _.map(found, 'match')
-  return found
+function action_seneca_close (args, done) {
+  this.emit('close')
+  done()
 }
 
 
-exports.inward = {
-  closed: inward_closed,
-  resolve_msg_id: inward_resolve_msg_id,
-  act_cache: inward_act_cache,
-  act_default: inward_act_default,
-  act_not_found: inward_act_not_found,
-  validate_msg: inward_validate_msg,
-  warnings: inward_warnings,
-  msg_meta: inward_msg_meta,
-  msg_modify: inward_msg_modify,
-  act_stats: inward_act_stats,
-  prepare_delegate: inward_prepare_delegate,
-  announce: inward_announce
-}
+function make_action_seneca_stats (private$) {
+  return function action_seneca_stats (args, done) {
+    args = args || {}
+    var stats
 
-
-exports.outward = {
-  act_cache: outward_act_cache,
-  res_object: outward_res_object,
-  act_stats: outward_act_stats
-}
-
-
-function inward_announce (ctxt, data) {
-  if (!ctxt.actmeta) return
-
-  if (_.isFunction(ctxt.seneca.on_act_in)) {
-    ctxt.seneca.on_act_in(ctxt.actmeta, data.msg)
-  }
-
-  ctxt.seneca.emit('act-in', data.msg)
-}
-
-
-function inward_closed (ctxt, data) {
-  if (ctxt.seneca.closed && !data.msg.closing$) {
-    return {
-      kind: 'error',
-      code: 'closed',
-      info: {
-        args: Util.inspect(Common.clean(data.msg)).replace(/\n/g, '')
-      }
+    if (args.pattern && private$.stats.actmap[args.pattern]) {
+      stats = private$.stats.actmap[args.pattern]
+      stats.time = private$.timestats.calculate(args.pattern)
     }
-  }
-}
+    else {
+      stats = _.clone(private$.stats)
+      stats.now = new Date()
+      stats.uptime = stats.now - stats.start
 
+      stats.now = new Date(stats.now).toISOString()
+      stats.start = new Date(stats.start).toISOString()
 
-function inward_resolve_msg_id (ctxt, data) {
-  var msg = data.msg
+      var summary =
+            (args.summary == null) ||
+            (/^false$/i.exec(args.summary) ? false : !!(args.summary))
 
-  var id_tx = (msg.id$ ||
-               msg.actid$ ||
-               msg.meta$.id ||
-               ctxt.seneca.idgen())
-        .split('/')
-
-  var tx =
-        id_tx[1] ||
-        msg.tx$ ||
-        msg.meta$.tx$ ||
-        ctxt.seneca.fixedargs.tx$ ||
-        ctxt.seneca.idgen()
-
-  var mi = (id_tx[0] || ctxt.seneca.idgen())
-
-  msg.meta$.mi = mi
-  msg.meta$.tx = tx
-  msg.meta$.id = mi + '/' + tx
-
-  ctxt.seneca.fixedargs.tx$ = tx
-}
-
-
-function inward_act_stats (ctxt, data) {
-  if (!ctxt.actmeta) {
-    return
-  }
-
-  var private$ = ctxt.seneca.private$
-  ++private$.stats.act.calls
-
-  var pattern = ctxt.actmeta.pattern
-
-  var actstats = (private$.stats.actmap[pattern] =
-                  private$.stats.actmap[pattern] || {})
-
-
-  ++actstats.calls
-}
-
-
-function outward_act_stats (ctxt, data) {
-  if (!ctxt.actmeta || ctxt.cached$) {
-    return
-  }
-
-  var private$ = ctxt.seneca.private$
-  ++private$.stats.act.done
-
-  var msg = data.msg
-
-  if (msg &&
-      msg.meta$ &&
-      msg.meta$.prior &&
-      msg.meta$.prior.entry
-     ) {
-    private$.timestats.point(ctxt.duration, ctxt.actmeta.pattern)
-  }
-
-  var pattern = ctxt.actmeta.pattern
-
-  var actstats = (private$.stats.actmap[pattern] =
-                  private$.stats.actmap[pattern] || {})
-
-
-  if (data.err) {
-    private$.stats.act.fails++
-    ++actstats.fails
-  }
-  else {
-    ++actstats.done
-  }
-}
-
-
-function inward_act_default (ctxt, data) {
-  var so = ctxt.options
-  var msg = data.msg
-
-  // TODO: existence of pattern action needs own indicator flag
-  if (!ctxt.actmeta) {
-    var default$ = msg.default$ || (!so.strict.find ? {} : msg.default$)
-
-    if (_.isPlainObject(default$) || _.isArray(default$)) {
-      return {
-        kind: 'result',
-        result: default$,
-        log: {
-          level: 'debug',
-          data: {
-            kind: 'act',
-            case: 'DEFAULT'
-          }
-        }
+      if (summary) {
+        stats.actmap = void 0
+      }
+      else {
+        _.each(private$.stats.actmap, function (a, p) {
+          private$.stats.actmap[p].time = private$.timestats.calculate(p)
+        })
       }
     }
 
-    else if (null != default$) {
-      return {
-        kind: 'error',
-        code: 'act_default_bad',
-        info: {
-          args: Util.inspect(Common.clean(msg)).replace(/\n/g, ''),
-          xdefault: Util.inspect(default$)
-        }
-      }
+    if (done) {
+      done(null, stats)
     }
+    return stats
   }
 }
 
 
-function inward_act_not_found (ctxt, data) {
-  var so = ctxt.options
-  var msg = data.msg
+function action_options_get (args, done) {
+  var options = this.options()
 
-  if (!ctxt.actmeta) {
-    return {
-      kind: 'error',
-      code: 'act_not_found',
-      info: { args: Util.inspect(Common.clean(msg)).replace(/\n/g, '') },
-      log: {
-        level: so.trace.unknown ? 'warn' : 'debug',
-        data: {
-          kind: 'act',
-          case: 'UNKNOWN'
-        }
-      }
-    }
-  }
+  var base = args.base || null
+  var root = base ? (options[base] || {}) : options
+  var val = args.key ? root[args.key] : root
+
+  done(null, Common.copydata(val))
 }
 
-
-function inward_validate_msg (ctxt, data) {
-  var so = ctxt.options
-  var msg = data.msg
-
-  Assert(ctxt.actmeta)
-
-  if (!_.isFunction(ctxt.actmeta.validate)) {
-    return
-  }
-
-  var err = null
-
-  // FIX: this is assumed to be synchronous
-  // seneca-parambulator and seneca-joi need to be updated
-  ctxt.actmeta.validate(msg, function (verr) {
-    err = verr
-  })
-
-  if (err) {
-    return {
-      kind: 'error',
-      code: so.legacy.error_codes ? 'act_invalid_args' : 'act_invalid_msg',
-      info: {
-        pattern: ctxt.actmeta.pattern,
-        message: err.message,
-        msg: Common.clean(msg),
-        error: err
-      },
-      log: {
-        level: so.trace.invalid ? 'warn' : null,
-        data: {
-          kind: 'act',
-          case: 'INVALID'
-        }
-      }
-
-    }
-  }
-}
-
-
-// Check if actid has already been seen, and if action cache is active,
-// then provide cached result, if any. Return true in this case.
-function inward_act_cache (ctxt, data) {
-  var so = ctxt.options
-  var msg = data.msg
-
-  var actid = msg.meta$.id
-  var private$ = ctxt.seneca.private$
-
-  if (actid != null && so.actcache.active) {
-    var actdetails = private$.actcache.get(actid)
-
-    if (actdetails) {
-      private$.stats.act.cache++
-
-      var err = actdetails.result[0]
-      var res = actdetails.result[1]
-
-      var out = {
-        kind: err ? 'error' : 'result',
-        result: res || null,
-        error: err || null,
-        log: {
-          level: 'debug',
-          data: {
-            kind: 'act',
-            case: 'CACHE'
-          }
-        }
-      }
-
-      ctxt.cached$ = true
-
-      return out
-    }
-  }
-}
-
-
-// Store result in action cache
-function outward_act_cache (ctxt, data) {
-  var so = ctxt.options
-  var msg = data.msg
-  var res = data.res
-  var err = data.err
-
-  var actid = msg.meta$.id
-  var private$ = ctxt.seneca.private$
-
-  if (actid != null && so.actcache.active) {
-    private$.actcache.set(actid, {
-      result: [err, res],
-      actmeta: ctxt.actmeta,
-      when: Date.now()
-    })
-  }
-}
-
-
-function inward_warnings (ctxt, data) {
-  var so = ctxt.options
-  var msg = data.msg
-
-  Assert(ctxt.actmeta)
-
-  if (so.debug.deprecation && ctxt.actmeta.deprecate) {
-    ctxt.seneca.log.warn({
-      kind: 'act',
-      case: 'DEPRECATED',
-      msg: msg,
-      pattern: ctxt.actmeta.pattern,
-      notice: ctxt.actmeta.deprecate,
-      callpoint: ctxt.callpoint
-    })
-  }
-}
-
-
-function inward_msg_meta (ctxt, data) {
-  var msg = data.msg
-
-  Assert(ctxt.actmeta)
-
-  msg.meta$.pattern = ctxt.actmeta.pattern
-  msg.meta$.action = ctxt.actmeta.id
-  msg.meta$.plugin_name = ctxt.actmeta.plugin_name
-  msg.meta$.plugin_tag = ctxt.actmeta.plugin_tag
-
-  msg.meta$.prior = msg.meta$.prior || { chain: [], entry: true, depth: 0 }
-  msg.meta$.start = ctxt.start
-  msg.meta$.sync = ctxt.sync
-}
-
-
-function inward_msg_modify (ctxt, data) {
-  data.msg = _.extend(
-    data.msg,
-    ctxt.seneca.fixedargs,
-    {tx$: data.msg.meta$.tx}
-  )
-
-  // remove actid so that user manipulation of msg for subsequent use does
-  // not cause inadvertent hit on existing action
-  delete data.msg.id$
-  delete data.msg.actid$ // legacy alias
-}
-
-
-function inward_prepare_delegate (ctxt, data) {
-  Assert(data.reply)
-
-  data.reply = data.reply.bind(ctxt.seneca)
-  data.reply.seneca = ctxt.seneca
-
-  var reply = data.reply
-
-  ctxt.seneca.good = function good (out) {
-    reply(null, out)
-  }
-
-  ctxt.seneca.bad = function bad (err) {
-    reply(err)
-  }
-}
-
-
-function outward_res_object (ctxt, data) {
-  var so = ctxt.options
-  var msg = data.msg
-  var res = data.res
-  var err = data.err
-
-  var not_object =
-        err == null &&
-        res != null &&
-        !(_.isPlainObject(res) ||
-          _.isArray(res) ||
-          !!res.entity$ ||
-          !!res.force$)
-
-  var not_legacy =
-        !(msg.cmd === 'generate_id' ||
-          msg.note === true ||
-          msg.cmd === 'native' ||
-          msg.cmd === 'quickcode')
-
-  if (so.strict.result && not_legacy && not_object) {
-    return {
-      kind: 'error',
-      code: 'result_not_objarr',
-      info: {
-        pattern: ctxt.actmeta.pattern,
-        args: Util.inspect(Common.clean(msg)).replace(/\n/g, ''),
-        result: res
-      }
-    }
-  }
-}
