+++ lib/legacy.js
@@ -14,6 +14,16 @@ var internals = {
   })
 }
 
+// All methods here are DEPRECATED
+// To be marked as DEPRECATED in Seneca 4.x
+// To be REMOVED in Seneca 5.x
+
+
+exports.hasact = function hasact (pattern) {
+  return !!this.find(pattern, {exact: false})
+}
+
+
 exports.fail = function make_legacy_fail (so) {
   return function () {
     var argsarr = new Array(arguments.length)
